Move code into the skyblock.item package
